+++ tests/scans/test_scan_display.py
@@ -86,7 +86,7 @@ def find_data_start(lines, fields):
 def extract_data(lines, shape, col_sep="|"):
     """ extract text data as numpy array """
     _h, w = shape
-    arry = numpy.empty(shape, dtype=numpy.float)
+    arry = numpy.empty(shape, dtype=float)
 
     incr = 0
     for line in lines:
@@ -182,7 +182,7 @@ def test_fast_scan_display(session, scan_data_listener_process):
                 start_once=True,
             )
             self.event = gevent.event.Event()
-            self.channels.append(AcquisitionChannel("block_data", numpy.int, ()))
+            self.channels.append(AcquisitionChannel("block_data", int, ()))
             self.pending_trigger = 0
             self.chunk = chunk
 
@@ -203,7 +203,7 @@ def test_fast_scan_display(session, scan_data_listener_process):
             return True
 
         def reading(self):
-            data = numpy.arange(self.npoints, dtype=numpy.int)
+            data = numpy.arange(self.npoints, dtype=int)
             acq_npoint = 0
             chunk = self.chunk
             i = 0
